Using the Escape Rope shows letters or JP characters instead of the proper sprite on DMG and doesn't spin correctly on SGB
When using the Escape Rope on an original Game Boy, for a split second while the player is shooting up into the sky, the player will use ABCD sprites instead of keep spinning like they're supposed to. The Super Game Boy has a similar error where the player doesn't spin as they shoot up, which are two variants of the same bug.